% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wxgb.error.R
\name{eval_loss}
\alias{eval_loss}
\title{Weighted mean of log loss (binary cross-entropy loss) for complex survey data}
\usage{
eval_loss(preds, dtrain)
}
\arguments{
\item{preds}{Predicted binary values (\code{0/1}) from the model.}

\item{dtrain}{An object of \code{xgb.DMatrix} including \code{label}(response variable or binary outcome), \code{data},\code{weight}.}
}
\value{
A list of two components:
\itemize{
\item \code{metric}: A string labelled as \code{weighted.mean.logloss}.
\item \code{value} : A numeric value of weighted mean of log loss.
}
}
\description{
A function to evaluate XGBoost models by weighted mean log loss for logistic regression. This function estimates predicted
probabilities from the predicted binary outcomes and computes the weighted mean of log likelihood values for Bernoulli distribution.
This function is a customized error function for complex survey data.
}
\seealso{
\code{\link[xgboost:xgb.train]{xgboost::xgb.train()}} for relevant arguments and return values in detail.

\code{\link[=wXGBoost]{wXGBoost()}} for detailed information on implementation.

\code{\link[=optim_wxgb_para]{optim_wxgb_para()}} for detailed optimization and examples.
}
