% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wxgb.error.R
\name{evalerror_reg}
\alias{evalerror_reg}
\title{Weighted mean squared error for complex survey data}
\usage{
evalerror_reg(preds, dtrain)
}
\arguments{
\item{preds}{Predicted values from the model}

\item{dtrain}{An object of \code{xgboost::xgb.DMatrix} including \code{label}(response variable), \code{data}, and \code{weight}.}
}
\value{
A list of two components:
\itemize{
\item \code{metric}: A string labelled as \code{weighted.mse}.
\item \code{value} : A numeric value of weighted mean squared error (MSE).
}
}
\description{
A function to evaluate XGBoost models by weighted mean squared error for linear regression.
This function is a customized error function for complex survey data.
}
\seealso{
\code{\link[xgboost:xgb.train]{xgboost::xgb.train()}} for relevant arguments and return values in detail.

\code{\link[=wXGBoost]{wXGBoost()}} for detailed information on implementation.

\code{\link[=optim_wxgb_para]{optim_wxgb_para()}} for detailed optimization and examples.
}
