% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_data.R
\docType{data}
\name{nhanes2013_sbc}
\alias{nhanes2013_sbc}
\title{2013-2014 cycle of NHANES data}
\format{
An object of class \code{data.frame} with 1510 obs. of 66 variables including
\describe{
\item{HBP}{Binary response variable labeled as \code{1} for hypertension or high blood pressure. }
\item{SEQN}{Respondent sequence number.}
\item{SDDSVYR}{Data Release Number indicating survey year.}
\item{SDMVPSU}{Masked Variance Pseudo-PSU, \code{cluster}.}
\item{SDMVSTRA}{Masked Variance Pseudo-Stratum, \code{strata}.}
\item{WTSAF2YR}{Fasting Subsample 2-year MEC Weight, \code{weights}.}
}
}
\source{
\url{https://www.cdc.gov/nchs/nhanes/about_nhanes.htm}
}
\usage{
data(nhanes2013_sbc)
}
\description{
The data was extracted from The National Center for Health Statistics (NCHS)
in the Centers for Diseases Control and Prevention (CDC) and has been processed to a sparse matrix for memory efficiency.
}
\examples{
data(nhanes2013_sbc)
}
\keyword{data}
