% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.weights.R
\name{replicate_weights}
\alias{replicate_weights}
\title{Replicate weights methods}
\usage{
replicate_weights(
  data,
  method = c("JKn", "dCV", "bootstrap", "subbootstrap", "BRR", "split", "extrapolation"),
  cluster = NULL,
  strata = NULL,
  weights = NULL,
  design = NULL,
  k = 10,
  R = 1,
  B = 200,
  train.prob = 0.7,
  method.split = c("dCV", "bootstrap", "subbootstrap"),
  rw.test = FALSE,
  dCV.sw.test = FALSE
)
}
\arguments{
\item{data}{A data frame with information on (at least) cluster and strata indicators, and sampling weights. It could be \code{NULL} if the sampling design were indicated in the \code{design} argument (see \code{design}).}

\item{method}{A string of a name of replicate weights methods. Choose one of these: \code{JKn}, \code{dCV}, \code{bootstrap}, \code{subbootstrap}, \code{BRR}, \code{split}, \code{extrapolation}.}

\item{cluster}{A string of a cluster identifier. It could be \code{NULL} if the sampling design were plugged in the \code{design} argument.}

\item{strata}{A string of a strata identifier. It could be \code{NULL} if the sampling design were plugged in the \code{design} argument.}

\item{weights}{A string of a sampling weights' identifier. It could be \code{NULL} if the sampling design were included in the \code{design} argument.}

\item{design}{An object of class \code{survey.design} generated by \code{\link[survey]{svydesign}}. It could be \code{NULL} if information about \code{cluster}, \code{strata}, \code{weights} and \code{data} were given.}

\item{k}{An integer. The number of folds to be defined. Default is \code{k=10}. Only applies to the \code{dCV} method.}

\item{R}{An integer. The number of times the sample is partitioned. Default is \code{R=1}. Only applies to \code{dCV}, \code{split} or \code{extrapolation} methods.}

\item{B}{An integer. The number of bootstrap re-samples. Default is \code{B=200}. Only applies to \code{bootstrap} and  \code{subbootstrap} methods.}

\item{train.prob}{A numeric between 0 and 1 that sets the proportion of clusters (for the method \code{split}) or strata (for the method \code{extrapolation}) in training sets. Default is \code{train.prob=0.7}. Only applies to \code{split} and \code{extrapolation} methods.}

\item{method.split}{A string of one the following replicate weights methods to be implemented under the \code{split} method in the \code{method} argument: \code{dCV}, \code{bootstrap} or \code{subbootstrap}.}

\item{rw.test}{A logical value. If \code{TRUE}, the function returns in the output object the replicate weights to the corresponding test sets. If \code{FALSE}, only the replicate weights of the training sets are returned. Default is \code{rw.test = FALSE}.}

\item{dCV.sw.test}{A logical value. If \code{TRUE}, original sampling weights for the units in the test sets are returned instead of the replicate weights. Default is \code{dCV.sw.test = FALSE}. Only applies to \code{dCV} method.}
}
\value{
The function returns a new data frame with new columns, each of which indicates replicate weights for different subsets.
}
\description{
A function calculating replicate weights.
}
\details{
This function from R-\code{wlasso} (\url{https://github.com/aiparragirre/wlasso/}, recently registered in CRAN: \code{svyVarSel})
is slightly modified for our package development.  The following description is also obtained from the original source for convenience.

Some of replicate weights methods, such as \code{JKn}, \code{bootstrap}, \code{subbootstrap} and \code{BRR},
are implemented in R-\code{survey}, \code{\link[survey]{as.svrepdesign}}, from which
the \code{replicate_weights()} function generates replicate weights for these options;
\code{dCV}, \code{split} and \code{extrapolation} have been expressly defined in this function.

Selecting any of the above-mentioned methods returns a new data frame,
in which new columns are added to the original data set; each column by replicate
weights always creates different pairs of training sets and optionally the corresponding test sets by setting \code{rw.test=TRUE}.
The number of new columns and their corresponding names are usually determined by numeric arguments, i.e.,\code{k}, or \code{R}, etc.
and specifically by the selected replicate weights method. The new columns indicating training and test sets
follow the similar structure for any selected method: \code{rw_r_x_train_t}, where \code{x=1,...,R}, and \code{t=1,...,T}
indicate the \code{x}-th partition of the sample,the \code{t}-th training set, respectively.
Similarly, the new columns for test sets are structured by \code{rw_r_x_test_t} or \code{sw_r_x_test_t}.
For other methods, \code{fold} or \code{set} for each unit respectively labeled as \code{fold_t} or \code{set_t}
is included in each partition depending on the method in a data set.
}
\examples{
data(nhanes2013_sbc)

## JKn ---------------------------------------------------------------------
data_rw <- replicate_weights(data = nhanes2013_sbc,
                             method = "JKn",
                             cluster = "SDMVPSU",
                             strata = "SDMVSTRA",
                             weights = "WTSAF2YR",
                             rw.test = TRUE)

## dCV ---------------------------------------------------------------------
data_rw <- replicate_weights(data = nhanes2013_sbc,
                             method = "dCV",
                             cluster = "SDMVPSU",
                             strata = "SDMVSTRA",
                             weights = "WTSAF2YR",
                             k = 10, R = 20,
                             rw.test = TRUE)

## subbootstrap ------------------------------------------------------------
data_rw <- replicate_weights(data = nhanes2013_sbc,
                             method = "subbootstrap",
                             cluster = "SDMVPSU",
                             strata = "SDMVSTRA",
                             weights = "WTSAF2YR",
                             B = 100)

## BRR ---------------------------------------------------------------------
data_rw <- replicate_weights(data = nhanes2013_sbc,
                             method = "BRR",
                             cluster = "SDMVPSU",
                             strata = "SDMVSTRA",
                             weights = "WTSAF2YR",
                             rw.test = TRUE)

## split ---------------------------------------------------------------------
\dontrun{
data_rw <- replicate_weights(data = nhanes2013_sbc,
                             method = "split",
                             cluster = "SDMVPSU",
                             strata = "SDMVSTRA",
                             weights = "WTSAF2YR",
                             R=20,
                             train.prob = 0.5,
                             method.split = "subbootstrap",
                             rw.test = TRUE)
}
## extrapolation -------------------------------------------------------------
data_rw <- replicate_weights(data = nhanes2013_sbc,
                            method = "extrapolation",
                            cluster = "SDMVPSU",
                            strata = "SDMVSTRA",
                            weights = "WTSAF2YR",
                            R=20,
                            train.prob = 0.5,
                            rw.test = TRUE)


}
