% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.wxgb.para.R
\name{generate_wxgb_params}
\alias{generate_wxgb_params}
\title{Generate randomized wXGBoost hyperparameters for tuning}
\usage{
generate_wxgb_params(custom_space = list())
}
\arguments{
\item{custom_space}{A named list of values or functions to override defaults. The list can contain static values (e.g., \code{subsample = 0.8})
or anonymous functions that return a single value when called (e.g., \code{function() sample(3:10, 1)}).
Default \code{objective} is \code{binary:logistic} for binary classification; \code{reg:squarederror} is set for linear regression.
The default search space is defined as follows:
\preformatted{
   defaults <- list(
     # General Parameters
     objective           = "binary:logistic",
     nthread             = 1,
     # Tree Booster Hyperparameters
     eta                 = runif(1, 0.01, 0.3),
     gamma               = runif(1, 0.0, 10.0),
     max_depth           = sample(3:10, 1),
     min_child_weight    = sample(1:20, 1),
     max_delta_step      = 0,               # for very imbalanced case : (1:10) helps
     subsample           = runif(1, 0.5, 1.0),
     colsample_bytree    = runif(1, 0.5, 1.0),
     colsample_bylevel   = 1.0,
     colsample_bynode    = 1.0,
     alpha               = runif(1, 0.0, 1.0), # L1 regularization
     lambda              = runif(1, 1.0, 4.0), # L2 regularization
     tree_method         = "auto",
     scale_pos_weight    = 1.0,
     # Random Seed
     seed                = sample.int(100000, 1)
     )
 }}
}
\value{
A named list of parameters ready to be parsed by \code{wXGBoost()}.
}
\description{
A helper function creating a complete list of valid wXGBoost hyperparameters.
It combines core structural configurations, regularization parameters, and
tree-building criteria. By default, it sets up a diverse random search space,
but users can inject fixed overrides or custom sampling closures.
}
\note{
The function is encapsulated in \code{optim_wxgb_para()} to dynamically create a (hyper)parameter set
for \code{wXGBoost()} per iteration for optimzation.
This function can also be used for \code{\link[xgboost:xgb.train]{xgboost::xgb.train()}} or \code{\link[xgboost:xgb.cv]{xgboost::xgb.cv()}}.
}
\examples{
# Generate using default search spaces
para.default<-generate_wxgb_params()

# Force a static max_depth and pass a custom eta sampler
custom.para<- generate_xgb_params(list(
  max_depth = 6,
  eta = function() runif(1, 0.05, 0.15)
))

}
